# Add cache for mmap IO