--- a/programs/local/LocalServer.cpp
+++ b/programs/local/LocalServer.cpp
@@ -260,6 +260,11 @@ try
     if (mark_cache_size)
         global_context->setMarkCache(mark_cache_size);
 
+    /// A cache for mmapped files.
+    size_t mmap_cache_size = config().getUInt64("mmap_cache_size", 1000);   /// The choice of default is arbitrary.
+    if (mmap_cache_size)
+        global_context->setMappedFileCache(mmap_cache_size);
+
     /// Load global settings from default_profile and system_profile.
     global_context->setDefaultProfiles(config());