--- a/programs/local/LocalServer.cpp
+++ b/programs/local/LocalServer.cpp
@@ -263,7 +263,7 @@ try
     /// A cache for mmapped files.
     size_t mmap_cache_size = config().getUInt64("mmap_cache_size", 1000);   /// The choice of default is arbitrary.
     if (mmap_cache_size)
-        global_context->setMappedFileCache(mmap_cache_size);
+        global_context->setMMappedFileCache(mmap_cache_size);
 
     /// Load global settings from default_profile and system_profile.
     global_context->setDefaultProfiles(config());