--- a/programs/server/Server.cpp
+++ b/programs/server/Server.cpp
@@ -829,6 +829,11 @@ int Server::main(const std::vector<std::string> & /*args*/)
     }
     global_context->setMarkCache(mark_cache_size);
 
+    /// A cache for mmapped files.
+    size_t mmap_cache_size = config().getUInt64("mmap_cache_size", 1000);   /// The choice of default is arbitrary.
+    if (mmap_cache_size)
+        global_context->setMappedFileCache(mmap_cache_size);
+
 #if USE_EMBEDDED_COMPILER
     size_t compiled_expression_cache_size = config().getUInt64("compiled_expression_cache_size", 500);
     CompiledExpressionCacheFactory::instance().init(compiled_expression_cache_size);