--- a/src/Common/CurrentMetrics.cpp
+++ b/src/Common/CurrentMetrics.cpp
@@ -66,6 +66,8 @@
     M(PartsWide, "Wide parts.") \
     M(PartsCompact, "Compact parts.") \
     M(PartsInMemory, "In-memory parts.") \
+    M(MappedFiles, "Total number of mmapped files.") \
+    M(MappedFileBytes, "Sum size of mmapped file regions.") \
 
 namespace CurrentMetrics
 {