--- a/src/Disks/DiskLocal.h
+++ b/src/Disks/DiskLocal.h
@@ -76,7 +76,8 @@ public:
         size_t buf_size,
         size_t estimated_size,
         size_t aio_threshold,
-        size_t mmap_threshold) const override;
+        size_t mmap_threshold,
+        MappedFileCache * mmap_cache) const override;
 
     std::unique_ptr<WriteBufferFromFileBase> writeFile(
         const String & path,