--- a/src/Disks/DiskMemory.h
+++ b/src/Disks/DiskMemory.h
@@ -67,7 +67,8 @@ public:
         size_t buf_size,
         size_t estimated_size,
         size_t aio_threshold,
-        size_t mmap_threshold) const override;
+        size_t mmap_threshold,
+        MappedFileCache * mmap_cache) const override;
 
     std::unique_ptr<WriteBufferFromFileBase> writeFile(
         const String & path,