--- a/src/Disks/DiskMemory.h
+++ b/src/Disks/DiskMemory.h
@@ -68,7 +68,7 @@ public:
         size_t estimated_size,
         size_t aio_threshold,
         size_t mmap_threshold,
-        MappedFileCache * mmap_cache) const override;
+        MMappedFileCache * mmap_cache) const override;
 
     std::unique_ptr<WriteBufferFromFileBase> writeFile(
         const String & path,