--- a/src/IO/MMapReadBufferFromFile.cpp
+++ b/src/IO/MMapReadBufferFromFile.cpp
@@ -44,7 +44,7 @@ MMapReadBufferFromFile::MMapReadBufferFromFile(const std::string & file_name_, s
     : file_name(file_name_)
 {
     open();
-    init(fd, offset, length_);
+    mapped.set(fd, offset, length_);
 }
 
 
@@ -52,7 +52,7 @@ MMapReadBufferFromFile::MMapReadBufferFromFile(const std::string & file_name_, s
     : file_name(file_name_)
 {
     open();
-    init(fd, offset);
+    mapped.set(fd, offset);
 }