@alexey-milovidov 2021-03-26
```
clickhouse-benchmark --checksum_on_read 0 --min_bytes_to_use_mmap_io 1 --query "SELECT sum(WatchID) FROM test.hits_none"
```

Before:
```
Queries executed: 437.

localhost:9000, queries 437, QPS: 139.723, RPS: 1239891245.364, MiB/s: 9459.619, result RPS: 139.723, result MiB/s: 0.001.

0.000%          0.006 sec.
10.000%         0.007 sec.
20.000%         0.007 sec.
30.000%         0.007 sec.
40.000%         0.007 sec.
50.000%         0.007 sec.
```

After:

```
localhost:9000, queries 5604, QPS: 210.707, RPS: 1869789962.386, MiB/s: 14265.365, result RPS: 210.707, result MiB/s: 0.002.

0.000%          0.004 sec.
10.000%         0.004 sec.
20.000%         0.004 sec.
30.000%         0.005 sec.
40.000%         0.005 sec.
50.000%         0.005 sec.
```