@alexey-milovidov 2021-03-28
Performance test has shown the expected improvement.